refactoring the conversions from editor state to JSON and vice versa
